+++ python/syntaqlite/__init__.py
@@ -34,3 +34,5 @@ def main():
         sys.exit(subprocess.call([binary] + sys.argv[1:]))
     else:
         os.execvp(binary, [binary] + sys.argv[1:])
+
+